Update _header.html.erb
The store and donate buttons on the navbar were not looking good as they were having unequal widths and were not aligned well. So I have made them align to right and also have made their widths equal. 

Along with that, the "About us" link on the navigation bar was aligned more towards the right which was not looking good. So I have made the margin-left of the "About us" link equal to the margin-left of other navigation links.